Fix inheritence when no preferred version status
When a version has a preferred child for reverse status inheritence, the
reverse inheritence routines did not handle a case where an issue did
not have a status set for the preferred child.  The fix allows the
version_status() function to fall back to using the last-status-found
method in these cases.